eleventh sprint project: transport catalogue

+++ transport-catalogue/domain.cpp
@@ -6,7 +6,7 @@ detail::Stop::Stop(std::string_view name, double lat, double lng)
     stop_crd.lng = lng;
 }
 
-info::BusInfo::BusInfo(std::string_view name, size_t stops, size_t u_stops, int route_length, double route_curvature)
+info::BusInfo::BusInfo(std::string_view name, size_t stops, size_t u_stops, double route_length, double route_curvature)
     :bus_name_info(name)
     ,stops_on_route(stops)
     ,unique_stops_on_route(u_stops)
